Move sub commands under cmd/kuma/cmd
